test: added in unit tests for all signals serialization/deserialization
This unit test uses automatically generated test data based on the attributes presented
in the signals, and then passes it through the serialization/deserialization process
to check whether the same output is obtained.

Currently this test will skip all signals that are not for use with the event bus, which
are signals that contain a list or dictionary record since they cannot be serialized yet.