feat!: Only accept metadata object in `send_event_with_custom_metadata` (#179)
This changes `send_event_with_custom_metadata` to accept a single
`metadata` object rather than all the fields that go into one.

It turns out that in practice, the event-bus-kafka consumer already needs
to build an EventsMetadata object or something equivalent to it as it
reconstructs the metadata from Kafka message headers, since that's where
the parsing and validation logic already is. It then has to splat the
object back out when calling this method. This leads to a possible
collision risk in the method's kwargs, but it's also just awkward. It is
likely that the same issue would arise in any other consumer.

Addresses https://github.com/openedx/openedx-events/issues/178