feat: Add configurable loading of event bus consumer (#200)
This defines an API for IDAs to use when consuming events that allows
the choice of Event Bus implementation to be configured rather than
written in code. It also includes a management command to start a
worker and allows to pass in extra arguments for the underlying
implementation of event bus. (`openedx_events` becomes a Django app
so that the management command can be registered.)

Closes https://github.com/openedx/openedx-events/issues/147

Co-authored-by: Tim McCormack <tmccormack@edx.org>