feat: add namespace to schemas (#212)
Add namespace to event schemas. This should be able to be deployed the producer and consumer in any order.
This will make the schema for a topic look like

{
 "doc": "Avro Event Format for CloudEvents created with openedx_events/schema",
  "fields": [<no change>],
  "name": "CloudEvent",
  "namespace": "org.openedx.content_authoring.course.catalog_info.changed.v1",
  "type": "record"
}
Adding the namespace ensures that there is no conflict when registering schemas by topic/record name, which is how we will be allowing more than one event type per topic. It will also make it easier to identify schemas for use in Avro unions if other users want to try that route.