fixup! Clarify language around send/publish/produce
"Send" was ambiguous because you can send to a Django signal and you can
send to the event bus, and one can lead to the other. New language:

- "Send" is used only for Django signals
- "Publish" is used for emitting events to the event bus broker

This also changes "producer mode" to "publishing mode".

I'm avoiding "produce" for the most part, as it leads to awkward phrasing
(in my opinion). Produce/consume and publish/subscribe are largely used as
synonyms in this space, but publish/subscribe seems to be more accurate. I
still want a way to describe IDAs and requests that can create an event
and try to publish it, with the distinction that the publishing may not
actually occur in all situations (e.g. in an error situation). For those,
"event-producing IDA" and "event-producing request" seem sufficient.