feat: Add autodoc for event annotations (#318)
* feat: Add autodoc for event annotations

* feat: Add event_key_field annotations where possible

Not all events are compatible with the bus, and some that are have never been configured for it that I
could find. This is everything I could locate in the
org.

* chore: Update to new code-annotations release and make upgrade

* chore: Finish removal of pii checks

It looks like the pii check config was removed at some point, but checks that pointed to it were left in place. This just finishes the removal.