feat: Raise exception when consume_indefinitely called without config
It's misleading to have the management command exit immediately with exit
code 0 when the config is missing. It's also a little harder than it should
be to debug this scenario because at most you get a warning log message
(which is easy to miss). A stack trace and error exit should be clearer.