feat: add signal that is triggered after a course is re-run
Adds `COURSE_RERUN_COMPLETED` signal which is similar to
`COURSE_IMPORT_COMPLETED` signal. It is triggered after a course is
re-run, allowing us to run some post re-run processes.

The current use of this is to run indexing for search and creation of
upstream links for components imported from libraries v2.